move all filesystem access behind abstraction
