+++ src/handlers/script/index.ts
@@ -1,6 +1,5 @@
 import { ipcMain } from 'electron'
 import log from 'electron-log/main'
-import { readFile, writeFile, unlink } from 'fs/promises'
 import path from 'path'
 
 import { SCRIPTS_PATH } from '@/constants/workspace'
@@ -9,6 +8,7 @@ import { showScriptSelectDialog, runScript } from '@/main/script'
 import { trackEvent } from '@/services/usageTracking'
 import { UsageEventName } from '@/services/usageTracking/types'
 import { browserWindowFromEvent } from '@/utils/electron'
+import { readFile, unlink, writeFile } from '@/utils/fs'
 import { ArchiveError, K6Client } from '@/utils/k6/client'
 import { TestRun } from '@/utils/k6/testRun'
 import { isExternalScript } from '@/utils/workspace'
