+++ src/main/script.ts
@@ -2,7 +2,6 @@ import { dialog, BrowserWindow } from 'electron'
 import log from 'electron-log/main'
 import type { Options } from 'k6/options'
 import { ChildProcessWithoutNullStreams } from 'node:child_process'
-import { createWriteStream } from 'node:fs'
 import { fileURLToPath, pathToFileURL } from 'node:url'
 import path from 'path'
 import * as tar from 'tar-stream'
@@ -12,6 +11,7 @@ import { TEMP_K6_ARCHIVE_PATH, TEMP_SCRIPT_SUFFIX } from '@/constants/workspace'
 import { ScriptHandler } from '@/handlers/script/types'
 import { getProxyArguments } from '@/main/proxy'
 import { ProxySettings } from '@/types/settings'
+import { createWriteStream } from '@/utils/fs'
 import { K6Client } from '@/utils/k6/client'
 import { createTrackingServer } from '@/utils/k6/tracking'
 import { readResource } from '@/utils/resources'
